Bug: Integer parameters for obtaining photos were formatted as doubles. (This mean we ended up within things like: "..?id=123.0"

+++ src/com/ml4d/ohow/MomentArrayAdapter.java
@@ -141,7 +141,7 @@ public class MomentArrayAdapter implements ListAdapter {
 			if (moment.getHasPhoto()) {
 				url = OHOWAPIResponseHandler.getBaseUrlIncludingTrailingSlash(false) + "photo.php"
 					+ "?" 
-					+ "id=" + Double.toString(moment.getId())
+					+ "id=" + Integer.toString(moment.getId())
 					+ "&photo_size=thumbnail";
 			} else {
 				url = null;
